+++ x/finality/keeper/power_dist_change.go
@@ -203,6 +203,10 @@ func (k Keeper) ProcessAllPowerDistUpdateEvents(
 					fpBTCPKHex := fpBTCPK.MarshalHex()
 					activeBTCDels[fpBTCPKHex] = append(activeBTCDels[fpBTCPKHex], btcDel)
 				}
+				// Becomes active, it should withdraw the rewards available
+				// and create a new BTCDelegationRewardsInfo.
+				// TODO: research how can we charge the gas for this process during
+				// the withdraw of rewards.
 			} else if delEvent.NewState == types.BTCDelegationStatus_UNBONDED {
 				// emit expired event if it is not early unbonding
 				if !btcDel.IsUnbondedEarly() {
@@ -210,6 +214,9 @@ func (k Keeper) ProcessAllPowerDistUpdateEvents(
 				}
 				// add the unbonded BTC delegation to the map
 				unbondedBTCDels[delEvent.StakingTxHash] = struct{}{}
+				// Becomes unbonded, it should withdraw the rewards available
+				// and create a new BTCDelegationRewardsInfo or delete it if does not have more sats
+				// for this finality provider.
 			}
 		case *types.EventPowerDistUpdate_SlashedFp:
 			// record slashed fps
@@ -269,6 +276,11 @@ func (k Keeper) ProcessAllPowerDistUpdateEvents(
 			btcDel := *dc.FinalityProviders[i].BtcDels[j]
 			if _, ok := unbondedBTCDels[btcDel.StakingTxHash]; !ok {
 				fp.AddBTCDelDistInfo(&btcDel)
+
+				err := k.IncentiveKeeper.BtcDelegationUnbonded(ctx, fp.GetAddress(), sdk.MustAccAddressFromBech32(btcDel.StakerAddr), btcDel.TotalSat)
+				if err != nil {
+					panic(err) // check if it should panic
+				}
 			}
 		}
 
@@ -318,6 +330,11 @@ func (k Keeper) ProcessAllPowerDistUpdateEvents(
 		fpActiveBTCDels := activeBTCDels[fpBTCPKHex]
 		for _, d := range fpActiveBTCDels {
 			fpDistInfo.AddBTCDel(d)
+
+			err := k.IncentiveKeeper.BtcDelegationActivated(ctx, sdk.MustAccAddressFromBech32(newFP.Addr), sdk.MustAccAddressFromBech32(d.StakerAddr), d.TotalSat)
+			if err != nil {
+				panic(err) // check if it should panic
+			}
 		}
 
 		// add this finality provider to the new cache if it has voting power